+++ lib/component/app.dart
@@ -35,8 +35,22 @@ class Arcane {
       Navigator.pop(context, result);
 
   static Future<T?> push<T extends Object?>(
-          BuildContext context, Widget child) =>
-      Pylon.push(context, child,
+          BuildContext context, Widget child) => 
+      Pylon.push<T?>(context, child,
+          type: PylonRouteType.material,
+          settings:
+              RouteSettings(name: child is ArcaneRoute ? child.path : null));
+  
+  static Future<T?> pushReplacement<T extends Object?, TO extends Object?>(
+          BuildContext context, Widget child) => 
+      Pylon.pushReplacement<T?, TO?>(context, child,
+          type: PylonRouteType.material,
+          settings:
+              RouteSettings(name: child is ArcaneRoute ? child.path : null));
+  
+  static Future<T?> pushAndRemoveUntil<T extends Object?>(
+          BuildContext context, Widget child, RoutePredicate predicate) => 
+      Pylon.pushAndRemoveUntil<T?>(context, child, predicate: predicate,
           type: PylonRouteType.material,
           settings:
               RouteSettings(name: child is ArcaneRoute ? child.path : null));