+++ lib/component/card_carousel.dart
@@ -1,10 +1,19 @@
 import 'package:arcane/arcane.dart';
 
-class CardCarousel extends StatefulWidget {
-  final List<Widget> children;
+class CardCarouselTheme {
   final int sharpness;
 
-  const CardCarousel({super.key, this.children = const [], this.sharpness = 9});
+  const CardCarouselTheme({this.sharpness = 9});
+
+  CardCarouselTheme copyWith({int? sharpness}) =>
+      CardCarouselTheme(sharpness: sharpness ?? this.sharpness);
+}
+
+class CardCarousel extends StatefulWidget {
+  final List<Widget> children;
+  final int? sharpness;
+
+  const CardCarousel({super.key, this.children = const [], this.sharpness});
 
   @override
   State<CardCarousel> createState() => _CardCarouselState();
@@ -48,7 +57,13 @@ class _CardCarouselState extends State<CardCarousel> {
                     decoration: BoxDecoration(
                         gradient: LinearGradient(colors: [
                       if (_controller.position.pixels > 0) cs.background,
-                      ...List.generate(widget.sharpness.clamp(1, 12).toInt(),
+                      ...List.generate(
+                          (widget.sharpness ??
+                                  ArcaneTheme.of(context)
+                                      .cardCarousel
+                                      .sharpness)
+                              .clamp(1, 12)
+                              .toInt(),
                           (_) => cs.background.withOpacity(0)),
                       if (_controller.hasClients &&
                           _controller.position.hasContentDimensions &&